requirements.txt
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
Flask==3.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kafka-python==2.1.5
MarkupSafe==3.0.2
oauthlib==3.2.2
praw==7.8.1
prawcore==2.4.0
python-dotenv==1.1.0
requests==2.32.3
requests-oauthlib==2.0.0
tweepy==4.15.0
update-checker==0.18.0
urllib3==2.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
